Removed some more functions from functions.json
