+++ scipy/special/xsf/bessel.h
@@ -1,7 +1,11 @@
 #pragma once
 
 #include "amos.h"
+#include "cephes/i0.h"
+#include "cephes/i1.h"
 #include "cephes/jv.h"
+#include "cephes/k0.h"
+#include "cephes/k1.h"
 #include "cephes/scipy_iv.h"
 #include "cephes/yv.h"
 #include "error.h"
@@ -633,6 +637,14 @@ T cyl_bessel_je(T v, T x) {
     return std::real(cyl_bessel_je(v, std::complex(x)));
 }
 
+inline double cyl_bessel_y0(double x) { return cephes::y0(x); }
+
+inline float cyl_bessel_y0(float x) { return cyl_bessel_y0(static_cast<double>(x)); }
+
+inline double cyl_bessel_y1(double x) { return cephes::y1(x); }
+
+inline float cyl_bessel_y1(float x) { return cyl_bessel_y1(static_cast<double>(x)); }
+
 inline std::complex<double> cyl_bessel_ye(double v, std::complex<double> z) {
     int n = 1;
     int kode = 2;
@@ -844,6 +856,14 @@ inline std::complex<float> cyl_hankel_2e(float v, std::complex<float> z) {
     );
 }
 
+inline double cyl_bessel_j0(double x) { return cephes::j0(x); }
+
+inline float cyl_bessel_j0(float x) { return cyl_bessel_j0(static_cast<double>(x)); }
+
+inline double cyl_bessel_j1(double x) { return cephes::j1(x); }
+
+inline float cyl_bessel_j1(float x) { return cyl_bessel_j1(static_cast<double>(x)); }
+
 inline std::complex<double> cyl_bessel_j(double v, std::complex<double> z) {
     int n = 1;
     int kode = 1;
@@ -973,6 +993,22 @@ inline double cyl_bessel_i(double v, double x) { return cephes::iv(v, x); }
 
 inline float cyl_bessel_i(float v, float x) { return cyl_bessel_i(static_cast<double>(v), static_cast<double>(x)); }
 
+inline double cyl_bessel_i0(double x) { return cephes::i0(x); }
+
+inline float cyl_bessel_i0(float x) { return cyl_bessel_i0(static_cast<double>(x)); }
+
+inline double cyl_bessel_i0e(double x) { return cephes::i0e(x); }
+
+inline float cyl_bessel_i0e(float x) { return cyl_bessel_i0e(static_cast<double>(x)); }
+
+inline double cyl_bessel_i1(double x) { return cephes::i1(x); }
+
+inline float cyl_bessel_i1(float x) { return cyl_bessel_i1(static_cast<double>(x)); }
+
+inline double cyl_bessel_i1e(double x) { return cephes::i1e(x); }
+
+inline float cyl_bessel_i1e(float x) { return cyl_bessel_i1e(static_cast<double>(x)); }
+
 inline std::complex<double> cyl_bessel_i(double v, std::complex<double> z) {
     int n = 1;
     int kode = 1;
@@ -1074,6 +1110,22 @@ T cyl_bessel_k(T v, T z) {
 
     return std::real(cyl_bessel_k(v, std::complex(z)));
 }
+
+inline double cyl_bessel_k0(double x) { return cephes::k0(x); }
+
+inline float cyl_bessel_k0(float x) { return cyl_bessel_k0(static_cast<double>(x)); }
+
+inline double cyl_bessel_k0e(double x) { return cephes::k0e(x); }
+
+inline float cyl_bessel_k0e(float x) { return cyl_bessel_k0e(static_cast<double>(x)); }
+
+inline double cyl_bessel_k1(double x) { return cephes::k1(x); }
+
+inline float cyl_bessel_k1(float x) { return cyl_bessel_k1(static_cast<double>(x)); }
+
+inline double cyl_bessel_k1e(double x) { return cephes::k1e(x); }
+
+inline float cyl_bessel_k1e(float x) { return cyl_bessel_k1e(static_cast<double>(x)); }
 
 inline std::complex<double> cyl_hankel_1(double v, std::complex<double> z) {
     int n = 1;
