+++ scipy/special/xsf/zeta.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "cephes/zeta.h"
+#include "cephes/zetac.h"
 
 namespace xsf {
 
@@ -10,11 +11,18 @@ XSF_HOST_DEVICE T zeta(T x, T q) {
 }
 
 template <>
-XSF_HOST_DEVICE inline float zeta(float xf, float qf) {
-    double x = xf;
-    double q = qf;
+XSF_HOST_DEVICE inline float zeta(float x, float q) {
+    return zeta(static_cast<double>(x), static_cast<double>(q));
+}
 
-    return zeta(x, q);
+template <typename T>
+XSF_HOST_DEVICE T zetac(T x) {
+    return cephes::zetac(x);
+}
+
+template <>
+XSF_HOST_DEVICE inline float zetac(float x) {
+    return zetac(static_cast<double>(x));
 }
 
 } // namespace xsf
